#81 explicitely used __all__ and move examples to the namespace
