+++ src/qmworks/data/dictionaries/__init__.py
@@ -1,4 +1,2 @@
 
-from .warningsCP2K import cp2k_warnings
-
-__all__ = ['cp2k_warnings']
+from .warningsCP2K import *
