+++ src/qmworks/data/dictionaries/warningsCP2K.py
@@ -1,3 +1,5 @@
+__all__ = ['cp2k_warnings']
+
 from qmworks.warnings_qmworks import SCF_Convergence_Warning
 
 cp2k_warnings = {
