+++ src/qmworks/examples/Conditional_workflows/__init__.py
@@ -1,0 +1,1 @@
+from .calc_freqs import example_freqs
