+++ src/qmworks/hdf5/__init__.py
@@ -1,5 +1,2 @@
-from .hdf5_functions import dump_to_hdf5
-from .quantumHDF5 import (StoreasHDF5, cp2k2hdf5, turbomole2hdf5)
-
-__all__ = ['StoreasHDF5', 'cp2k2hdf5', 'dump_to_hdf5',
-           'turbomole2hdf5']
+from .hdf5_functions import *
+from .quantumHDF5 import *
