+++ src/qmworks/packages/__init__.py
@@ -1,10 +1,6 @@
-from .packages import (Package, run, registry, Result,
-                       SerMolecule, SerSettings)
-from .cp2k_package import cp2k
-from .SCM import (adf, dftb)
-from .orca import orca
-from .gamess import gamess
-from .dirac import dirac
-
-__all__ = ['Package', 'Result', 'SerMolecule', 'SerSettings', 'adf', 'cp2k',
-           'dftb', 'dirac', 'gamess', 'orca', 'registry', 'run']
+from .packages import *
+from .cp2k_package import *
+from .SCM import *
+from .orca import *
+from .gamess import *
+from .dirac import *
