+++ src/qmworks/parsers/__init__.py
@@ -1,15 +1,8 @@
-from .adf_parser import  (extract_properties_rkf, kfreader)
-from .cp2KParser import (readCp2KBasis, read_cp2k_coefficients, readCp2KOverlap,
-                         read_cp2k_number_of_orbitals)
-from .dirac_xml_parser import parse_xml
+from .adf_parser import *
+from .cp2KParser import *
+from .dirac_xml_parser import *
 from .gamess_parser import *
-from  .generic_parsers import awk_file
+from  .generic_parsers import *
 from .orca_parser import *
-from .turbomoleParser import (readTurbomoleBasis, readTurbomoleMO)
-from .xyzParser import(manyXYZ, parse_string_xyz, readXYZ, string_to_plams_Molecule)
-
-
-# __all__ = ['manyXYZ', 'parse_string_xyz', 'readCp2KBasis',
-#            'readCp2KCoeff',
-#            'readCp2KOverlap', 'readTurbomoleBasis', 'readTurbomoleMO',
-#            'readXYZ', 'read_cp2k_number_of_orbitals']
+from .turbomoleParser import *
+from .xyzParser import *
