+++ src/qmworks/warnings_qmworks.py
@@ -1,3 +1,4 @@
+__all__ = ['SCF_Convergence_Warning', 'Geometry_Convergence_Warning']
 
 
 class SCF_Convergence_Warning(RuntimeWarning):
